Program now recording and merging file data

+++ Emlin/ConstantValues.cs
@@ -12,7 +12,7 @@ namespace Emlin
     {
         public const int NUMBER_OF_INPUTS = 128;
         public const int NUMBER_OF_COOMBINATIONS = NUMBER_OF_INPUTS * NUMBER_OF_INPUTS;
-        public const double LENGTH_OF_SESSION_IN_MILLIS = 1500;
+        public const int LENGTH_OF_SESSION_IN_MILLIS = 1500;
         public static string KEYBOARD_DATA_FILEPATH = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) , Assembly.GetCallingAssembly().GetName().Name);
     }
 }
